requirements.txt
streamlit
pandas
python-dateutil
streamlit-authenticator
fpdf2
python-dateutil